move game to shared
